Refactor project architecture
