+++ src/database/model/User.ts
@@ -1,0 +1,1 @@
+import { Schema } from "mongoose";
