+++ src/graphql/resolvers/plants.ts
